Move Auth type up to http package
